pyproject.toml
[project]
name = "001-doc-extractor-platform"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "celery>=5.6.2",
    "fastapi>=0.135.1",
    "langchain>=1.2.10",
    "langchain-anthropic>=1.3.4",
    "langchain-openai>=1.1.10",
    "langchain-text-splitters>=1.1.1",
    "langgraph>=1.0.10",
    "langsmith>=0.7.11",
    "openai>=2.24.0",
    "psycopg2-binary>=2.9.11",
    "pymupdf>=1.27.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "qdrant-client>=1.17.0",
    "redis>=7.2.1",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
]
